:bug: pytest error due to side effects
- running order matters...
  - `test_errors.py` fails if `test_models.py`
     is run before. I don't know fastai or other might change in the state.
  - running only `test_errors.py` works fine, running it before as well.

- made some adoptions to other tests which clutter the commit a bit.
